style: File structuring
