+++ src/Components/Interfaces/projectItemProps.ts
@@ -1,0 +1,5 @@
+import { Project } from './projects';
+
+export interface ProjectItemProps {
+    project: Project;
+  }
