+++ src/Components/Interfaces/projects.ts
@@ -1,0 +1,8 @@
+export interface Project {
+    image: string;
+    title: string;
+    languages: string;
+    description: string;
+    github?: string;
+    website?: string;
+  }
