+++ src/Components/Projects/Projects.tsx
@@ -1,7 +1,7 @@
 // src/components/Projects.tsx
 import React from 'react';
 import ProjectItem from './ProjectItem';
-import projects from '../Data/portfolioProjects';
+import { projects } from '../Data/portfolioProjects';
 import './Projects.scss';
 
 const Projects: React.FC = () => {
